+++ packages/nocodb/package.json
@@ -21,8 +21,8 @@
     "local:test:graphql": "cross-env DATABASE_URL=mysql://root:password@localhost:3306/sakila TS_NODE_PROJECT=tsconfig.json mocha -r ts-node/register   src/__tests__/graphql.test.ts --recursive --timeout 10000  --exit",
     "test:graphql": "cross-env TS_NODE_PROJECT=tsconfig.json mocha -r ts-node/register   src/__tests__/graphql.test.ts --recursive --timeout 10000  --exit",
     "test:grpc": "cross-env TS_NODE_PROJECT=tsconfig.json mocha -r ts-node/register   src/__tests__/grpc.test.ts --recursive --timeout 10000  --exit",
-    "local:test:rest": "cross-env TS_NODE_PROJECT=./src/__tests__/tsconfig.json mocha -r ts-node/register   src/__tests__/unit/rest/index.test.ts --recursive --timeout 50000  --exit",
-    "test:rest": "cross-env TS_NODE_PROJECT=./src/__tests__/tsconfig.json mocha -r ts-node/register   src/__tests__/unit/rest/index.test.ts --recursive --timeout 50000  --exit",
+    "local:test:rest": "cross-env TS_NODE_PROJECT=./tests/unit/tsconfig.json mocha -r ts-node/register   tests/unit/rest/index.test.ts --recursive --timeout 50000  --exit",
+    "test:rest": "cross-env TS_NODE_PROJECT=./tests/unit/tsconfig.json mocha -r ts-node/register   tests/unit/rest/index.test.ts --recursive --timeout 50000  --exit",
     "test1": "run-s build test:*",
     "test:lint": "tslint --project . && prettier \"src/**/*.ts\" --list-different",
     "test:unit": "nyc --silent ava",
